Move app/cogs/type to app/types
