+++ app/types/interaction.go
