+++ go.mod
@@ -6,5 +6,4 @@ require (
 	github.com/bwmarrin/discordgo v0.24.0
 	github.com/rs/zerolog v1.26.1
 	github.com/spf13/viper v1.11.0
-	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
